read시 결과값 배열로 반환
readPost 및 readPostList의 결과값을 Array<PostData> 타입으로 반환합니다.
