+++ BoostcampRelay23/BoostcampRelay23/SceneDelegate.swift
@@ -19,7 +19,9 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
         // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
         // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-
+        
+        
+        
         // Create the SwiftUI view that provides the window contents.
         let contentView = ContentView()
